Updated to work with gcc using `-Wall -Werror`
* gcc seems pickier than clang, in a good way.
* Threw my hands in the air and realized that bringing in GLAD makes more sense
  than reintroducing function pointers all over the place in the engine code.
* BUG: Game starts but QC bytecode interpretation fails hard (segfault). Not
  sure if this is a Win32 thing or because of the code updates (note to self:
  see if this happens on macOS/clang).

+++ src/net_main.c
@@ -424,7 +424,7 @@ NET_CheckNewConnections
 struct {
     double time;
     int op;
-    long session;
+    uintptr_t session;
 } vcrConnect;
 
 qsocket_t *NET_CheckNewConnections(void) {
@@ -444,7 +444,7 @@ qsocket_t *NET_CheckNewConnections(void) {
             if(recording) {
                 vcrConnect.time = host_time;
                 vcrConnect.op = VCR_OP_CONNECT;
-                vcrConnect.session = (long)ret;
+                vcrConnect.session = (uintptr_t)ret;
                 Sys_FileWrite(vcrFile, &vcrConnect, sizeof(vcrConnect));
                 Sys_FileWrite(vcrFile, ret->address, NET_NAMELEN);
             }
@@ -499,7 +499,7 @@ returns -1 if connection is invalid
 struct {
     double time;
     int op;
-    long session;
+    uintptr_t session;
     int ret;
     int len;
 } vcrGetMessage;
@@ -543,7 +543,7 @@ int NET_GetMessage(qsocket_t *sock) {
         if(recording) {
             vcrGetMessage.time = host_time;
             vcrGetMessage.op = VCR_OP_GETMESSAGE;
-            vcrGetMessage.session = (long)sock;
+            vcrGetMessage.session = (uintptr_t)sock;
             vcrGetMessage.ret = ret;
             vcrGetMessage.len = net_message.cursize;
             Sys_FileWrite(vcrFile, &vcrGetMessage, 24);
@@ -553,7 +553,7 @@ int NET_GetMessage(qsocket_t *sock) {
         if(recording) {
             vcrGetMessage.time = host_time;
             vcrGetMessage.op = VCR_OP_GETMESSAGE;
-            vcrGetMessage.session = (long)sock;
+            vcrGetMessage.session = (uintptr_t)sock;
             vcrGetMessage.ret = ret;
             Sys_FileWrite(vcrFile, &vcrGetMessage, 20);
         }
@@ -576,7 +576,7 @@ returns -1 if the connection died
 struct {
     double time;
     int op;
-    long session;
+    uintptr_t session;
     int r;
 } vcrSendMessage;
 
@@ -601,7 +601,7 @@ int NET_SendMessage(qsocket_t *sock, sizebuf_t *data) {
     if(recording) {
         vcrSendMessage.time = host_time;
         vcrSendMessage.op = VCR_OP_SENDMESSAGE;
-        vcrSendMessage.session = (long)sock;
+        vcrSendMessage.session = (uintptr_t)sock;
         vcrSendMessage.r = r;
         Sys_FileWrite(vcrFile, &vcrSendMessage, 20);
     }
@@ -630,7 +630,7 @@ int NET_SendUnreliableMessage(qsocket_t *sock, sizebuf_t *data) {
     if(recording) {
         vcrSendMessage.time = host_time;
         vcrSendMessage.op = VCR_OP_SENDMESSAGE;
-        vcrSendMessage.session = (long)sock;
+        vcrSendMessage.session = (uintptr_t)sock;
         vcrSendMessage.r = r;
         Sys_FileWrite(vcrFile, &vcrSendMessage, 20);
     }
@@ -664,7 +664,7 @@ qboolean NET_CanSendMessage(qsocket_t *sock) {
     if(recording) {
         vcrSendMessage.time = host_time;
         vcrSendMessage.op = VCR_OP_CANSENDMESSAGE;
-        vcrSendMessage.session = (long)sock;
+        vcrSendMessage.session = (uintptr_t)sock;
         vcrSendMessage.r = r;
         Sys_FileWrite(vcrFile, &vcrSendMessage, 20);
     }
